feat(TheBarberHub-2/Sprints#8): Añadir Capa Dominio Back

+++ tienda/src/main/java/com/fpmislata/daw/tienda/persistence/dao/jpa/GenericJpaDao.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 public interface GenericJpaDao<T> {
     List<T> findAll(int page, int size);
 
-    Optional<T> findById(int id);
+    Optional<T> findById(long id);
 
     T insert(T entity);
